proc: Fix Binary.End in execve map during proc parsing
The following policy

apiVersion: cilium.io/v1alpha1
kind: TracingPolicy
metadata:
  name: "test"
spec:
  kprobes:
  - call: "security_bprm_check"
    syscall: false
    return: true
    args:
    - index: 0
      type: "path"
      resolve: "file.f_path"
    returnArg:
      index: 0
      type: "int"
    returnArgAction: "Post"
    selectors:
    - matchBinaries:
      - operator: Postfix
        values:
        - "bash"

generates events only for bash processes that are created after Tetragon
starts. If the bash shell is started before Tetragon, it fails to
generate events.

It seems that in struct execve_map_value we set bin->end only on eBPF
code (processes that start after Tetragon) but not in proc parsing
(processes that started before Tetragon).

This patch fixes that.
