+++ pkg/sensors/exec/procevents/proc_reader.go
@@ -18,7 +18,6 @@ import (
 	"github.com/cilium/tetragon/pkg/observer"
 	"github.com/cilium/tetragon/pkg/option"
 	"github.com/cilium/tetragon/pkg/reader/proc"
-	"github.com/cilium/tetragon/pkg/sensors/exec/execvemap"
 	"github.com/cilium/tetragon/pkg/sensors/exec/userinfo"
 )
 
@@ -244,41 +243,6 @@ func pushExecveEvents(p procs, inInitTreeMap map[uint32]struct{}) {
 
 		observer.AllListeners(&m)
 	}
-}
-
-func procToKeyValue(p procs, inInitTree map[uint32]struct{}) (*execvemap.ExecveKey, *execvemap.ExecveValue) {
-	k := &execvemap.ExecveKey{Pid: p.pid}
-	v := &execvemap.ExecveValue{}
-
-	v.Parent.Pid = p.ppid
-	v.Parent.Ktime = p.pktime
-	v.Process.Pid = p.pid
-	v.Process.Ktime = p.ktime
-	v.Flags = 0
-	v.Nspid = p.nspid
-	v.Capabilities.Permitted = p.permitted
-	v.Capabilities.Effective = p.effective
-	v.Capabilities.Inheritable = p.inheritable
-	v.Namespaces.UtsInum = p.utsNs
-	v.Namespaces.IpcInum = p.ipcNs
-	v.Namespaces.MntInum = p.mntNs
-	v.Namespaces.PidInum = p.pidNs
-	v.Namespaces.PidChildInum = p.pidForChildrenNs
-	v.Namespaces.NetInum = p.netNs
-	v.Namespaces.TimeInum = p.timeNs
-	v.Namespaces.TimeChildInum = p.timeForChildrenNs
-	v.Namespaces.CgroupInum = p.cgroupNs
-	v.Namespaces.UserInum = p.userNs
-	pathLength := copy(v.Binary.Path[:], p.exe)
-	v.Binary.PathLength = int32(pathLength)
-
-	_, parentInInitTree := inInitTree[p.ppid]
-	if v.Nspid == 1 || parentInInitTree {
-		v.Flags |= api.EventInInitTree
-		inInitTree[p.pid] = struct{}{}
-	}
-
-	return k, v
 }
 
 func pushEvents(ps []procs) {
